Apply background colour when drawing Box border
The borderStyle field only encodes the foreground colour of the border,
leaving the background to inherit from the terminal default theme. This
often means the border is draw white-on-white, which is a regression
introduced by the tcell v2 update.

It needs to honour the background colour set for the Box instance
when drawing the border.

Fixes: https://github.com/rivo/tview/issues/521
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>